Move files and update dependencies
