+++ packages/workflow/src/index.ts
@@ -4,6 +4,7 @@ import * as NodeHelpers from './NodeHelpers';
 import * as ObservableObject from './ObservableObject';
 import * as TelemetryHelpers from './TelemetryHelpers';
 
+export * from './AIParametersParser';
 export * from './errors';
 export * from './Constants';
 export * from './Cron';
